+++ src/components/Buttons.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+const Buttons = () => {
+    return(
+        <div className="btns">
+            <button 
+                className="btn reset-btn">
+                Reset</button>
+            <button 
+                className="btn numbers-btn">
+                Show me lucky numbers</button>
+        </div>
+    )
+}
+
+export default Buttons;
