+++ src/components/Headers.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+const Headers = () => {
+    return(
+        <div className="headers">
+            <h1>Lotto 6/49</h1>
+            <h3>Generating lucky numbers</h3>
+        </div>
+    )
+}
+
+export default Headers;
