+++ src/components/Numbers.js
@@ -1,0 +1,29 @@
+import React, {Component} from "react";
+
+
+const Numbers = () => {
+    return (
+        <div className="numbers">
+        <span>1</span>
+        <span>1</span>
+        <span>1</span>
+        <span>1</span>
+        <span>1</span>
+        <span>1</span>
+        <span className="special-num">1</span>
+    </div>
+    )
+}
+
+export default Numbers;
+/* 
+export default class Numbers extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+
+        };
+        
+    }
+    render()
+} */
