+++ src/index.js
@@ -1,6 +1,6 @@
 
 import React from "react"
-import App from "./App.jsx"
+import App from "./App.js"
 import reactDOM from "react-dom"
 
 reactDOM.render(<App/>,document.getElementById("root"))
